pyproject.toml
[tool.poetry]
name = "pythonproject5"
version = "0.1.0"
description = ""
authors = ["Notebook <zekserj@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
django-dotenv = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
